edited to read in list of queries from file, and perform twitter search and SA on each
