+++ twitter_utilities.py
@@ -5,6 +5,34 @@ from nltk.tag import pos_tag
 from nltk.stem.wordnet import WordNetLemmatizer
 
 
+def cleanup_query(search_query):
+    for symbol in ["@", "#", "$"]:
+        if symbol in search_query:
+            clean_query = search_query.replace(symbol, "")
+        else:
+            clean_query = search_query
+    return clean_query
+
+
+def get_api_keys(keyfile):
+    """Load API keys from local textfile"""
+    with open(keyfile, "r") as f:
+        file_data = f.readlines()
+    keys = {}
+    for line in file_data:
+        key_type, value = line.split(": ")
+        keys[key_type] = value.strip()
+    return keys
+
+
+def get_search_queries():
+    QUERY_FILE = "/Users/daniellefevre/PycharmProjects/untitled2/queries.txt"
+    with open(QUERY_FILE, "r") as f:
+        query_list = f.readlines()
+    query_list = [query.strip() for query in query_list]
+    return query_list
+
+
 def remove_noise(tweet_tokens, stop_words=()):
 
     cleaned_tokens = []
@@ -27,3 +55,4 @@ def remove_noise(tweet_tokens, stop_words=()):
         if len(token) > 0 and token not in string.punctuation and token.lower() not in stop_words:
             cleaned_tokens.append(token.lower())
     return cleaned_tokens
+
